myAppManager/requirements.txt
playsound~=1.3.0
customtkinter~=5.2.2
pillow~=10.4.0